Matrix Methods

This page contains methods for matrix operations.

MultivariateChebyshev.cartesian_product — Function

cartesian_product(arrays::AbstractVector{<:AbstractVector})

Calculate the cartesian product of a list of input arrays.

Parameter arrays represents a list of input vectors.

Method returns a Matrix of size (N, D) where D equals the number of input vectors and N equals the product of lengths of input vectors.

MultivariateChebyshev.matmul — Function

matmul(A::AbstractArray, B::AbstractArray)

Generalised matrix multiplication along last two dimensions.

This method mimics Numpy's matmul behaviour.

MultivariateChebyshev.batchmul — Function

batchmul(A::AbstractArray, B::AbstractArray)

Generalised matrix multiplication along first two dimensions.